+++ src/styles/DateInput.scss
@@ -1,0 +1,9 @@
+.DateInput {
+
+}
+
+@media screen and (min-width: $sm) {
+    .DateInput {
+        
+    }
+}
