+++ src/styles/index.scss
@@ -43,3 +43,4 @@ code {
 @import './Navigation.scss';
 @import './ProjectTimeTracking.scss';
 @import './RangeDatePicker.scss';
+@import './DateInput.scss';
